Update README.md
My project is very simple to run I download this project from git hub on 23 June 2024 on Sunday.  My project name is calculator . the project that I download from git hub has five classes that name is calculator and when we get the coding of calculator then in this coding just five functions are given but when I changing then in this coding I Add different three functions . in calculator wo solve different ISSUES.
for example in calculator we add the numbers we can subtract the number and we multiply the numbers we can modulus use in this  and I ADD the different features.
we cab factorial of many numbers and also solve the issues in our question therefore I CHOOSE this project of calculator and solve the issues then I upload the project on git hub.